Adopt createConditionalCommentHousekeeper in inlineAssetsPerQueryString also
