+++ lib/transforms/inlineAssetsPerQueryString.js
@@ -1,3 +1,5 @@
+const createConditionalCommentHousekeeper = require('../createConditionalCommentHousekeeper');
+
 // Inlines assets as data: urls
 //
 //   If the url of the image has an 'inline' GET parameter, it is inlined if the parameter has a true (or no) value, eg.: foo.png?inline or foo.png?inline=true
@@ -17,6 +19,8 @@ module.exports = (queryObj, { minimumIeVersion = Infinity } = {}) => {
 
   return function inlineAssetsPerQueryString(assetGraph) {
     const minimumIeVersionByAsset = new Map(); // Asset => minimumIeVersion
+    const getOpenConditionalComments = createConditionalCommentHousekeeper();
+
     function getMinimumIeVersionUsage(asset, stack = []) {
       if (asset.type === 'Html') {
         return 1;
@@ -34,13 +38,15 @@ module.exports = (queryObj, { minimumIeVersion = Infinity } = {}) => {
               matchCondition = incomingRelation.condition.match(
                 /^(gte?|lte?)\s+IE\s+(\d+)$/i
               );
-            } else if (
-              incomingRelation.conditionalComments &&
-              incomingRelation.conditionalComments.length > 0
-            ) {
-              matchCondition = incomingRelation.conditionalComments[0].nodeValue.match(
-                /^\[if\s+(gte?|lte?)\s+IE\s+(\d+)\s*\]\s*>\s*<\s*!\s*$/i
+            } else if (incomingRelation.from.type === 'Html') {
+              const openConditionalComments = getOpenConditionalComments(
+                incomingRelation
               );
+              if (openConditionalComments.length > 0) {
+                matchCondition = openConditionalComments[0].match(
+                  /^\[if\s+(gte?|lte?)\s+IE\s+(\d+)\s*\]\s*>\s*<\s*!\s*$/i
+                );
+              }
             }
             if (matchCondition) {
               if (matchCondition[1].substr(0, 2) === 'lt') {
